add missing authorization

+++ API/Controllers/LikeController.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using API.Data;
 using API.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -15,12 +17,16 @@ namespace API.Controllers
     public class LikeController : ControllerBase
     {
         private readonly AppDbContext context;
-        public LikeController(AppDbContext context)
+        private readonly UserManager<User> _userManager;
+
+        public LikeController(AppDbContext context, UserManager<User> userManager)
         {
             this.context = context;
+            _userManager = userManager;
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> LikeImage(int imageId)
         {
             var userId = User.Claims.SingleOrDefault(x => x.Type.Equals("http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier")).Value;
@@ -50,6 +56,7 @@ namespace API.Controllers
         }
 
         [HttpDelete()]
+        [Authorize]
         public async Task<IActionResult> DislikeImage(int imageId)
         {
             var userId = User.Claims.SingleOrDefault(x => x.Type.Equals("http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier")).Value;
@@ -61,8 +68,13 @@ namespace API.Controllers
                 return NotFound("Like not found.");
             }
 
+            if (like.UserId != userId && !await _userManager.IsInRoleAsync(await _userManager.FindByIdAsync(userId), "Moderator"))
+            {
+                return Unauthorized();
+            }
+
             context.Likes.Remove(like);
-            await context.SaveChangesAsync();
+            await context.SaveChangesAsync(); // TODO: test this. Saving once not enough?
 
             var image = await context.Images.FindAsync(imageId);
             image.LikeCount--;
